requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Marshmallow==3.20.1
Flask-Marshmallow==0.15.0
Marshmallow-SQLAlchemy==0.29.0
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1